Move SubscriptionSelection component to component folder
