+++ src/client/src/components/SubscriptionSelection/SubscriptionSelection.test.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import configureMockStore from "redux-mock-store";
 import SubscriptionSelection from "./SubscriptionSelection";
 import { Provider } from "react-redux";
-import { getInitialState, setSubscriptions } from "../../../mockData/mockStore";
+import { getInitialState, setSubscriptions } from "../../mockData/mockStore";
 import { render, RenderResult } from "@testing-library/react";
 import { IntlProvider } from "react-intl";
