+++ src/client/src/containers/AppServiceModal/index.tsx
@@ -10,7 +10,7 @@ import { isAppServiceModalOpenSelector } from "../../selectors/modalSelector";
 import RuntimeStackInfo from "./RuntimeStackInfo/RuntimeStackInfo";
 import AppServicePlanInfo from "./AppServicePlanInfo/AppServicePlanInfo";
 import AppNameEditor from "./AppNameEditor/AppNameEditor";
-import SubscriptionSelection from "./SubscriptionSelection/SubscriptionSelection";
+import SubscriptionSelection from "../../components/SubscriptionSelection/SubscriptionSelection";
 import { InjectedIntlProps, injectIntl } from "react-intl";
 import buttonStyles from "../../css/buttonStyles.module.css";
 import { WIZARD_CONTENT_INTERNAL_NAMES, KEY_EVENTS } from "../../utils/constants";
